v0.5.4 updates
support freeze, thaw, and frozen
include key images of spent outputs when txs created
bump supported monero-project to v0.17.2.0
add cryptonote_format_utils_basic dependency

+++ src/daemon/monero_daemon_model.h
@@ -192,7 +192,7 @@ namespace monero {
     boost::optional<uint64_t> m_weight;
     std::vector<std::shared_ptr<monero_output>> m_inputs;
     std::vector<std::shared_ptr<monero_output>> m_outputs;
-    std::vector<uint32_t> m_output_indices;
+    std::vector<uint64_t> m_output_indices;
     boost::optional<std::string> m_metadata;
     boost::optional<std::string> m_common_tx_sets;
     std::vector<uint8_t> m_extra;
